Split Postgres class into subfiles
